* run audio in the same process as engine — much easier audio clock sync, easy transition to multithreaded version in future * remove video module — will be replaced with SFML version

+++ src/oscillators.jl
@@ -29,5 +29,5 @@ function sine(ν::Signal{Float64}, θ::Function)
   sine(signal(ν), θ)
 end
 
-sine(ν, θ=0.0) = sine(constantly(ν), constantly(θ))
+sine(ν::Float64, θ=0.0) = sine(constantly(ν), constantly(θ))
 sine(ν::Function, θ=0.0) = sine(ν, constantly(θ))
